Squash all 4.2.2 changes into one
Bringing up the keyboard when a picture is the last item visible will always scroll the conversation to the bottom

Bug 7511504

The code to scroll needs to test that the tall item is the last item in
the list.

Change-Id: I4a9d72178564b8047c40af1dc148904b75e79694

Unable to create application com.android.mms.MmsApp

Bug 7597021

Handle null result from detectCountry. Let the onCountryDetected listener
do the work and delay setting mCountryIso. Tested getCurrentCountryIso()
by always returning null to see that the app didn't crash.

Change-Id: I452958d857ee9c651a3d552300f39a5e0e5b281f

Area code is not displayed in To: field, while sending SMS from call log

Bug 7614149

The extra space after the comma was screwing up the Chips parser and causing
the comma delimeter to be parsed as part of the number.

Change-Id: I6239f6b1e45075b6faf9e7f0323d00945b23ce73

SMS conversation list doesn't mantain scroll position

Bug 7585885

Change-Id: Ic500e0b3245b5f42a9e7994827a3825b041b1c01

Opening recent apps while on compose dismisses compose on re-rentering app

Bug 7586014

Don't bail back to ConversationList when we've got an empty ComposeMessageActivity.
In the case where the user has entered a recipient and followed the same
repro steps in the bug, a draft is created (see CMA's saveDraft) because
the receipient editor is open and there is a recipient. In that case, the
working message is not discarded and does not go through the escape-to-
ConversationList case. By the way, I was not able to repro the original
bug 2793874 that added this escape clause code.

Change-Id: I92d84f19c8ba2c27a8c28975dc95d1d230b89f1c

Fix Derp

Change-Id: Ie53502583436799b4292adc47a345eb935d666ba